+++ src/artist/artist.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
-import { ArtistRepository } from './artist.repository';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { ArtistEntity } from './entity/artist.entity';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([ArtistEntity])],
   controllers: [ArtistController],
-  providers: [ArtistService, ArtistRepository],
+  providers: [ArtistService],
 })
 export class ArtistModule {}
